% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annicchiarico.R
\name{annicchiarico}
\alias{annicchiarico}
\title{Annicchiarico (1992) method}
\usage{
annicchiarico(data, env, gen, rep, var)
}
\arguments{
\item{data}{Data file (data.frame) with variables.}

\item{env}{Column with environment information.}

\item{gen}{Coluna contendo informações de genótipo.}

\item{rep}{Column containing genotype information.}

\item{var}{Variable to be analyzed.}
}
\value{
A data frame containing the following estimates:
  \itemize{
    \item{\code{Gen}}{: Genotype.}
    \item{\code{Wg}}{: General recommendation index for environments.}
    \item{\code{Wd}}{: Recommendation index for unfavorable environments.}
    \item{\code{Wf}}{: Recommendation index for favorable environments.}
    \item{\code{GE}}{: Membership (\%) to the general stability genotypes group.}
    \item{\code{PA}}{: Membership (\%) to the  poorly adapted geotypes group.}
    \item{\code{FAV}}{: Membership (\%) to the favorable adaptade genotypes group.}
    \item{\code{UNF}}{: Membership (\%) to the unffavorable adaptade genotypes group.}
  }
}
\description{
Adaptability and Stability Analysis based on the interpretation of the Annicchiarico 1992 methodology, developed by Carneiro et al. 2019.
}
\examples{
# Caminho do arquivo de dados
caminho <- system.file("extdata", "Dados.csv", package = "safuzzy")

# Leitura do arquivo
dados = read.csv2(caminho)

# Uso da função com a variável de interesse
annicchiarico(data = dados, env = environment, gen = treatment, rep = block, var = gy)

}
\references{
Carneiro, A. R. T., Sanglard, D. A., Azevedo, A. M., Souza, T. L. P. O. D., Pereira, H. S., & Melo, L. C. (2019). Fuzzy logic in automation for interpretation of adaptability and stability in plant breeding studies. Scientia Agricola, 76(2), 123-129.
}
\seealso{
\code{\link{hybrid}}

\code{\link{lin_binns}}
}
\author{
Douglas de Oliveira Maciel \email{douglasmaciel@discente.ufg.br}
}
