% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruz_torres_venckovisk.R
\name{cruz_torres_vencovsky}
\alias{cruz_torres_vencovsky}
\title{Cruz, Torres & Vencovsky (1989) method}
\usage{
cruz_torres_vencovsky(data, env, gen, rep, var)
}
\arguments{
\item{data}{Data file (data.frame) with variables.}

\item{env}{Column with environment information.}

\item{gen}{Coluna contendo informações de genótipo.}

\item{rep}{Column containing genotype information.}

\item{var}{Variable to be analyzed.}
}
\value{
A data frame containing the following estimates:
  \itemize{
    \item{\code{Gen}}{: Genotype.}
    \item{\code{B_0}}{: Mean of the variable for each genotype.}
    \item{\code{B_1}}{: Regression coefficient (B_1) for each genotype.}
    \item{\code{B1_B2}}{: Regression coefficient (B1+B2) for each genotype.}
    \item{\code{R2}}{: Coefficient of determination (R²) for each genotype.}
    \item{\code{MdAF}}{: Membership (\%) to the average adaptability to favorable environments genotypes group.}
    \item{\code{Nda}}{: Membership (\%) to the  poorly adapted geotypes group.}
    \item{\code{MdAG}}{: Membership (\%) to the general adaptability to favorable environments genotypes group.}
    \item{\code{MaxGF}}{: Membership (\%) to the maximum adaptability to favorable environments genotypes group.}
    \item{\code{MaxDes}}{: Membership (\%) to the maximum adaptability to unfavorable environments genotypes group.}
    \item{\code{BE}}{: Membership (\%) to the low stability genotypes group.}
    \item{\code{BP}}{: Membership (\%) to the low yield genotypes group.}
  }
}
\description{
Adaptability and Stability Analysis based on the interpretation of the Cruz, Torres & Vencovsky (1989), developed by Carneiro et al. 2019.
}
\examples{
# Caminho do arquivo de dados
caminho <- system.file("extdata", "Dados.csv", package = "safuzzy")

# Leitura do arquivo
dados = read.csv2(caminho)

# Uso da função com a variável de interesse
cruz_torres_vencovsky(data = dados, env = environment, gen = treatment, rep = block, var = gy)
}
\references{
Carneiro, A. R. T., Sanglard, D. A., Azevedo, A. M., Souza, T. L. P. O. D., Pereira, H. S., & Melo, L. C. (2019). Fuzzy logic in automation for interpretation of adaptability and stability in plant breeding studies. Scientia Agricola, 76(2), 123-129.
}
\seealso{
\code{\link{hybrid}}

\code{\link{lin_binns}}

\code{\link{eberhart_russell}}
}
\author{
Douglas de Oliveira Maciel \email{douglasmaciel@discente.ufg.br}
}
