% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eberhart_russel.R
\name{eberhart_russel}
\alias{eberhart_russel}
\title{Eberhart & Russel método}
\usage{
eberhart_russel(dados, ambiente, genotipo, bloco, variavel)
}
\arguments{
\item{dados}{Arquivo de dados (data.frame) com as variáveis.}

\item{ambiente}{Coluna com as informações de embiente.}

\item{genotipo}{Coluna contendo informações de genótipo.}

\item{bloco}{coluna contendo informações de bloco.}

\item{variavel}{Variável a ser análisada.}
}
\value{
Um data frame contendo as seguintes estimativas:
  \itemize{
    \item{\code{Gen}}{: Genótipo.}
    \item{\code{B_0}}{: Média da variável para cada genótipo.}
    \item{\code{B_1}}{: Coeficiente de regressão (B_1i) para cada genótipo.}
    \item{\code{R2}}{: Coeficiente de determinação (R²) padronizado para cada genótipo.}
    \item{\code{GE}}{: Pertinência (\%) para o grupo de genótipos de estabilidade geral.}
    \item{\code{PA}}{: Pertinência (\%) para o grupo de genótipos pouco adaptados.}
    \item{\code{FAV}}{: Pertinência (\%) para o grupo de genótipos com adaptabilidade favorável.}
    \item{\code{Des}}{: Pertinência (\%) para o grupo de genótipos adaptados a ambientes desfavoráveis.}
  }
}
\description{
Análise de Adaptabilidade e Estabilidade baseado na interpretação da metodologia de Eberhart & Russel 1966, desenvolvido por Carneiro et al. 2018.
}
\examples{
# Caminho do arquivo de dados
caminho <- system.file("extdata", "Dados.csv", package = "safuzzy")

# Leitura do arquivo
dados = read.csv2(caminho)

# Uso da função com a variável de interesse
eberhart_russel(dados, ambiente = environment, genotipo = treatment, bloco = block, variavel = gy)

}
\references{
Carneiro, V. Q., Prado, A. L. D., Cruz, C. D., Carneiro, P. C. S., Nascimento, M., & Carneiro, J. E. D. S. (2018). Fuzzy control systems for decision-making in cultivars recommendation. Acta Scientiarum. Agronomy, 40, e39314.
}
\author{
Douglas de Oliveira Maciel \email{douglasmaciel@discente.ufg.br}
}
