% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hibrido.R
\name{hibrido}
\alias{hibrido}
\title{Híbrido método}
\usage{
hibrido(dados, ambiente, genotipo, bloco, variavel)
}
\arguments{
\item{dados}{Arquivo de dados (data.frame) com as variáveis.}

\item{ambiente}{Coluna com as informações de embiente.}

\item{genotipo}{Coluna contendo informações de genótipo.}

\item{bloco}{coluna contendo informações de bloco.}

\item{variavel}{Variável a ser análisada.}
}
\value{
Um data frame contendo as seguintes estimativas:
  \itemize{
    \item{\code{Gen}}{: Genótipo.}
    \item{\code{PIF}}{: Índice de performance em ambientes favoráveis.}
    \item{\code{PID}}{: Índice de performance em ambientes desfavoráveis.}
    \item{\code{B_1}}{: Coeficiente de regressão (B_1i) para cada genótipo.}
    \item{\code{R2}}{: Coeficiente de determinação (R²) padronizado para cada genótipo.}
    \item{\code{GE}}{: Pertinência (\%) para o grupo de genótipos de estabilidade geral.}
    \item{\code{PA}}{: Pertinência (\%) para o grupo de genótipos pouco adaptados.}
    \item{\code{FAV}}{: Pertinência (\%) para o grupo de genótipos com adaptabilidade favorável.}
    \item{\code{Des}}{: Pertinência (\%) para o grupo de genótipos adaptados a ambientes desfavoráveis.}
  }
}
\description{
Análise de Estabilidade e Adaptabilidade baseado na interpretação da metodologia de Eberhart & Russel 1966, Associado ao método de Lins & Bins modificado, desenvolvido por Carneiro et al. (2020).
}
\examples{
# Caminho do arquivo de dados
caminho <- system.file("extdata", "Dados.csv", package = "safuzzy")

# Leitura do arquivo
dados = read.csv2(caminho)

# Uso da função com a variável de interesse
hibrido(dados, ambiente = environment, genotipo = treatment, bloco = block, variavel = gy)

}
\references{
Carneiro, A. R. T., Sanglard, D. A., Azevedo, A. M., Souza, T. L. P. O. D., Pereira, H. S., Melo, L. C., & Carneiro, P. C. S. (2020). Fuzzy logic applied to different adaptability and stability methods in common bean. Pesquisa agropecuária brasileira, 55, e01609.
}
\author{
Douglas de Oliveira Maciel \email{douglasmaciel@discente.ufg.br}
}
