% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{hybrid}
\alias{hybrid}
\title{Hybrid method}
\usage{
hybrid(data, env, gen, rep, var)
}
\arguments{
\item{data}{Data file (data.frame) with variables.}

\item{env}{Column with environment information.}

\item{gen}{Coluna contendo informações de genótipo.}

\item{rep}{Column containing genotype information.}

\item{var}{Variable to be analyzed.}
}
\value{
Um data frame contendo as seguintes estimativas:
  \itemize{
    \item{\code{Gen}}{: Genotype.}
    \item{\code{PIF}}{: Performance index in favorable environments.}
    \item{\code{PID}}{: Performance index in unfavorable environments.}
    \item{\code{B_1}}{: Regression coefficient (B_1) for each genotype.}
    \item{\code{R2}}{: Standardized coefficient of determination (R²) for each genotype.}
    \item{\code{GE}}{: Membership (\%) to the general stability genotypes group.}
    \item{\code{PA}}{: Membership (\%) to the  poorly adapted geotypes group.}
    \item{\code{FAV}}{: Membership (\%) to the favorable adaptade genotypes group.}
    \item{\code{UNF}}{: Membership (\%) to the unffavorable adaptade genotypes group.}
  }
}
\description{
Stability and Adaptability Analysis based on the interpretation of the Eberhart & Russel 1966 methodology, associated with the modified Lins & Bins (1988) methodology, developed by Carneiro et al. (2020).
}
\examples{
# Caminho do arquivo de dados
caminho <- system.file("extdata", "Dados.csv", package = "safuzzy")

# Leitura do arquivo
dados = read.csv2(caminho)

# Uso da função com a variável de interesse
hybrid(data = dados, env = environment, gen = treatment, rep = block, var = gy)

}
\references{
Carneiro, A. R. T., Sanglard, D. A., Azevedo, A. M., Souza, T. L. P. O. D., Pereira, H. S., Melo, L. C., & Carneiro, P. C. S. (2020). Fuzzy logic applied to different adaptability and stability methods in common bean. Pesquisa agropecuária brasileira, 55, e01609.
}
\seealso{
\code{\link{eberhart_russell}}

\code{\link{lin_binns}}
}
\author{
Douglas de Oliveira Maciel \email{douglasmaciel@discente.ufg.br}
}
