% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_fuzzy.R
\name{eberhart_russel}
\alias{eberhart_russel}
\title{Eberhart & Russel método}
\usage{
eberhart_russel(Dados, variavel)
}
\arguments{
\item{Dados}{Arquivo de dados com as variáveis. Deve ser organizado na seguinte ordem: Ambiente, Genótipo, Bloco.}

\item{variavel}{Variável a ser análisada.}
}
\value{
Um data frame contendo as seguintes estimativas:
   \item{\code{Media(B_0)}}{Para cada genótipo.}
}
\description{
Análise de Estabilidade e Adaptabilidade baseado na interpretação da metodologia de Eberhart & Russel 1966, desenvolvido por Carneiro et al. 2018.
}
\examples{
eberhart_russel(Dados = dados, varivavel = prod)
}
\references{
Carneiro, V. Q., Prado, A. L. D., Cruz, C. D., Carneiro, P. C. S., Nascimento, M., & Carneiro, J. E. D. S. (2018). Fuzzy control systems for decision-making in cultivars recommendation. Acta Scientiarum. Agronomy, 40, e39314.
}
