% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_fuzzy.R
\name{hibrido}
\alias{hibrido}
\title{Híbrido método}
\usage{
hibrido(Dados, variavel)
}
\arguments{
\item{Dados}{Arquivo de dados com as variáveis. Deve ser organizado na seguinte ordem: Ambiente, Genótipo, Bloco.}

\item{variavel}{Variável a ser análisada.}
}
\description{
Análise de Estabilidade e Adaptabilidade baseado na interpretação da metodologia de Eberhart & Russel 1966, Associado ao método de Lins & Bins modificado, desenvolvido por Carneiro et al. (2020).
}
\references{
Carneiro, A. R. T., Sanglard, D. A., Azevedo, A. M., Souza, T. L. P. O. D., Pereira, H. S., Melo, L. C., & Carneiro, P. C. S. (2020). Fuzzy logic applied to different adaptability and stability methods in common bean. Pesquisa agropecuária brasileira, 55, e01609.
}
